% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_week_half.R
\name{iso_week_half}
\alias{iso_week_half}
\title{Week number of date, plus a half if midweek}
\usage{
iso_week_half(date)
}
\arguments{
\item{date}{a date}
}
\value{
a (double) week number, with extra half if Tuesday, Wednesday or Thursday
}
\description{
Week number of date, plus a half if midweek
}
\examples{
iso_week_half(as.Date("2020-07-03"))
iso_week_half(as.Date("2020-07-02"))

}
